Code for Invoice

+++ Project/AutoTourism/Source Code/AutoTourism/Lodge/AutoTourism.Lodge.Facade/RoomReservation/Dto.cs
@@ -11,9 +11,9 @@ namespace AutoTourism.Lodge.Facade.RoomReservation
     {        
         public DateTime BookingDate { get; set; }
         public DateTime BookingFrom { get; set; }
-        public Int16 NoOfDays { get; set; }
-        public Int16 NoOfPersons { get; set; }
-        public Int16 NoOfRooms { get; set; }
+        public Int32 NoOfDays { get; set; }
+        public Int32 NoOfPersons { get; set; }
+        public Int32 NoOfRooms { get; set; }
         public Double Advance { get; set; }
         public Int64 BookingStatusId { get; set; }
