+++ Project/Crystal/Source Code/Crystal/Lodge/Crystal.Lodge.Component/Room/Reservation/Data.cs
@@ -6,9 +6,9 @@ namespace Crystal.Lodge.Component.Room.Reservation
 
     public class Data : Crystal.Reservation.Component.Data
     {
-        public Int16 NoOfDays { get; set; }
-        public Int16 NoOfPersons { get; set; }
-        public Int16 NoOfRooms { get; set; }      
+        public Int32 NoOfDays { get; set; }
+        public Int32 NoOfPersons { get; set; }
+        public Int32 NoOfRooms { get; set; }      
         public String Description { get; set; }
         public Double Advance { get; set; }
         public Boolean IsCheckedIn { get; set; }
